Module Tjr_plist_freelist

A freelist, based on plist; don't open this module.

module Freelist_intf : sig ... end

Main interfaces used by the freelist

module Freelist_make : sig ... end

module Make_2 : sig ... end

Like Freelist_make but with std_types

module Make_3 : sig ... end

Like Make_2, but combined with plist construction.

module Freelist_example : sig ... end

include Freelist_intf

type ('blk_id, 't) root_block_ops = {

write_freelist_roots : hd:'blk_id -> tl:'blk_id -> (unit, 't) Tjr_monad.m;

sync : unit -> (unit, 't) Tjr_monad.m;

}

type sync_type = [

| `Tl_only

| `Tl_and_root_block

]

`Tl: this just ensures that the tl block of the freelist is synced; it does not update the root block

`Tl_and_root_block: sync tl and global root block

type fIXME

type ('blk_id, 't) freelist_ops = {

alloc : unit -> ('blk_id, 't) Tjr_monad.m;

alloc_many : int -> (fIXME, 't) Tjr_monad.m;

free : 'blk_id -> (unit, 't) Tjr_monad.m;

free_many : fIXME -> (unit, 't) Tjr_monad.m;

sync : sync_type -> (unit, 't) Tjr_monad.m;

}

alloc_many: int is the number of blk_ids (although this sort-of forces us to unmarshal the whole block; perhaps prefer also storing the number of elements in the block, precisely for this reason FIXME

free_many: free an entire list of blk_ids, by appending to this freelist

type 'a min_free_ops = {

min_free_alloc : 'a -> int -> 'a list * 'a;

}

type 'a freelist = {

transient : 'a list;

min_free : ('a * 'a min_free_ops) option;

waiting : 'a Tjr_monad.event list;

disk_thread_active : bool;

}

The freelist state, in addition to the plist.

transient: a non-persistent list of free elts that can be allocated

min_free: free elts are stored in transient, or in the on-disk plist; usually we also have an "upper bound", beyond which we can still allocate elts freely (an elt is either in transient, allocated to some object, in the plist, or geq min_free

waiting: a list of threads waiting for a disk process to return

NOTE: we assume that the state is accessed via with_state, ie, only one thread at a time

val make : < async : (unit -> (unit, 't) Tjr_monad.m) -> (unit, 't) Tjr_monad.m; event_ops : 't Tjr_monad.event_ops; monad_ops : 't Tjr_monad.monad_ops; plist : ('elt, 'buf, 'blk_id, 't) Tjr_plist.Plist_intf.plist_ops; root_block : ('blk_id, 't) root_block_ops; version : ('elt, 'blk_id, 't) Freelist_make.version; with_freelist : ('elt freelist, 't) Tjr_monad.with_state; > -> ('elt, 't) freelist_ops

let make : 
< async : (unit -> (unit, 't) m) -> (unit, 't) m;
  event_ops : 't event_ops;
  monad_ops : 't monad_ops;
  plist : ('elt, 'buf, 'blk_id, 't) plist_ops;
  root_block : ('blk_id, 't) root_block_ops;
  version : ('elt, 'blk_id, 't) version;
  with_freelist : ('elt freelist, 't) with_state;
> -> ('elt, 't) freelist_ops
= make