Module Fl_example_1.With_

Parameters

S : sig ... end

Signature

val plist_ops : Tjr_plist.Pl_type_abbrevs.blk_id Tjr_plist.Plist_intf.Pl_origin.pl_origin -> ((Tjr_fs_shared.Shared_ctxt.r, Tjr_fs_shared.ba_buf, Tjr_plist.Pl_type_abbrevs.blk_id, Tjr_plist.Pl_type_abbrevs.t) Tjr_plist.Plist_intf.plist_ops, Tjr_monad.With_lwt.lwt) Tjr_monad.m

val min_free_alloc : Tjr_fs_shared.Shared_ctxt.r -> int -> Tjr_fs_shared.Shared_ctxt.B.blk_id Tjr_lib.List_.t * Tjr_fs_shared.Shared_ctxt.B.blk_id

val with_plist_ops : (Tjr_fs_shared.Shared_ctxt.r, 'a, Tjr_fs_shared.Shared_ctxt.r, Tjr_monad.With_lwt.lwt) Tjr_plist.Plist_intf.plist_ops -> < with_locked_ref : Tjr_fs_shared.Shared_ctxt.r Tjr_plist_freelist.Freelist_intf.freelist_im -> < freelist_ops : (Tjr_fs_shared.Shared_ctxt.r, Tjr_fs_shared.Shared_ctxt.r, Tjr_monad.With_lwt.lwt) Tjr_plist_freelist.Freelist_intf.freelist_ops; freelist_ref : Tjr_fs_shared.Shared_ctxt.r Tjr_plist_freelist.Freelist_intf.freelist_im Stdlib.ref; >; with_state : (Tjr_fs_shared.Shared_ctxt.r Tjr_plist_freelist.Freelist_intf.freelist_im, Tjr_monad.With_lwt.lwt) Tjr_monad.with_state -> (Tjr_fs_shared.Shared_ctxt.r, Tjr_fs_shared.Shared_ctxt.r, Tjr_monad.With_lwt.lwt) Tjr_plist_freelist.Freelist_intf.freelist_ops; >

val wrap : sync_origin:(('c, 'd) Tjr_plist_freelist.Freelist_intf.Fl_origin.t -> (unit, Tjr_monad.With_lwt.lwt) Tjr_monad.m) -> freelist_ops:('c, 'd, Tjr_monad.With_lwt.lwt) Tjr_plist_freelist.Freelist_intf.freelist_ops -> ('a -> ('b, Tjr_fs_shared.Shared_ctxt.t) Tjr_monad.m) -> 'a -> ('b, Tjr_monad.With_lwt.lwt) Tjr_monad.m

This wraps the freelist operations in an inefficient piece of code that checks whether the origin has changed, and if it has, it syncs the origin

val add_origin_autosync : blk_id:Tjr_fs_shared.Shared_ctxt.r -> freelist_ops:(Tjr_fs_shared.Shared_ctxt.r, Tjr_fs_shared.Shared_ctxt.r, Tjr_monad.With_lwt.lwt) Tjr_plist_freelist.Freelist_intf.freelist_ops -> (Tjr_fs_shared.Shared_ctxt.r, Tjr_fs_shared.Shared_ctxt.r, Tjr_monad.With_lwt.lwt) Tjr_plist_freelist.Freelist_intf.freelist_ops

val from_origin : Tjr_fs_shared.Shared_ctxt.r -> (< freelist_ops : (Tjr_fs_shared.Shared_ctxt.r, Tjr_fs_shared.Shared_ctxt.r, Tjr_monad.With_lwt.lwt) Tjr_plist_freelist.Freelist_intf.freelist_ops; freelist_ref : Tjr_fs_shared.Shared_ctxt.r Tjr_plist_freelist.Freelist_intf.freelist_im Stdlib.ref; >, Tjr_monad.With_lwt.lwt) Tjr_monad.m

val from_origin_with_autosync : Tjr_fs_shared.Shared_ctxt.r -> (< freelist_ops : (Tjr_fs_shared.Shared_ctxt.r, Tjr_fs_shared.Shared_ctxt.r, Tjr_monad.With_lwt.lwt) Tjr_plist_freelist.Freelist_intf.freelist_ops; freelist_ref : Tjr_fs_shared.Shared_ctxt.r Tjr_plist_freelist.Freelist_intf.freelist_im Stdlib.ref; >, Tjr_monad.With_lwt.lwt) Tjr_monad.m