Module Freelist_intf.Fl_origin

type ('a, 'blk_id) fl_origin = {

hd : 'blk_id;

tl : 'blk_id;

blk_len : int;

min_free : 'a option;

}

val bin_shape_fl_origin : Bin_prot.Shape.t -> Bin_prot.Shape.t -> Bin_prot.Shape.t

val bin_size_fl_origin : a blk_id. 'a Bin_prot.Size.sizer -> 'blk_id Bin_prot.Size.sizer -> ('a, 'blk_id) fl_origin Bin_prot.Size.sizer

val bin_write_fl_origin : a blk_id. 'a Bin_prot.Write.writer -> 'blk_id Bin_prot.Write.writer -> ('a, 'blk_id) fl_origin Bin_prot.Write.writer

val bin_writer_fl_origin : 'a Bin_prot.Type_class.writer -> 'b Bin_prot.Type_class.writer -> ('a, 'b) fl_origin Bin_prot.Type_class.writer

val __bin_read_fl_origin__ : a blk_id. 'a Bin_prot.Read.reader -> 'blk_id Bin_prot.Read.reader -> (int -> ('a, 'blk_id) fl_origin) Bin_prot.Read.reader

val bin_read_fl_origin : a blk_id. 'a Bin_prot.Read.reader -> 'blk_id Bin_prot.Read.reader -> ('a, 'blk_id) fl_origin Bin_prot.Read.reader

val bin_reader_fl_origin : 'a Bin_prot.Type_class.reader -> 'b Bin_prot.Type_class.reader -> ('a, 'b) fl_origin Bin_prot.Type_class.reader

val bin_fl_origin : 'a Bin_prot.Type_class.t -> 'b Bin_prot.Type_class.t -> ('a, 'b) fl_origin Bin_prot.Type_class.t

type ('a, 'blk_id) t = ('a, 'blk_id) fl_origin

val bin_shape_t : Bin_prot.Shape.t -> Bin_prot.Shape.t -> Bin_prot.Shape.t

val bin_size_t : a blk_id. 'a Bin_prot.Size.sizer -> 'blk_id Bin_prot.Size.sizer -> ('a, 'blk_id) t Bin_prot.Size.sizer

val bin_write_t : a blk_id. 'a Bin_prot.Write.writer -> 'blk_id Bin_prot.Write.writer -> ('a, 'blk_id) t Bin_prot.Write.writer

val bin_writer_t : 'a Bin_prot.Type_class.writer -> 'b Bin_prot.Type_class.writer -> ('a, 'b) t Bin_prot.Type_class.writer

val __bin_read_t__ : a blk_id. 'a Bin_prot.Read.reader -> 'blk_id Bin_prot.Read.reader -> (int -> ('a, 'blk_id) t) Bin_prot.Read.reader

val bin_read_t : a blk_id. 'a Bin_prot.Read.reader -> 'blk_id Bin_prot.Read.reader -> ('a, 'blk_id) t Bin_prot.Read.reader

val bin_reader_t : 'a Bin_prot.Type_class.reader -> 'b Bin_prot.Type_class.reader -> ('a, 'b) t Bin_prot.Type_class.reader

val bin_t : 'a Bin_prot.Type_class.t -> 'b Bin_prot.Type_class.t -> ('a, 'b) t Bin_prot.Type_class.t

type ('a, 'blk_id, 't) ops = {

read : unit -> (('a, 'blk_id) t, 't) Tjr_monad.m;

write : ('a, 'blk_id) t -> (unit, 't) Tjr_monad.m;

}

NOTE barrier and sync only make sense if the non-origin part of the freelist is on the same blk dev