Module Tjr_plist_freelist__Summary

Summary, for documentation purposes

include Tjr_plist_freelist.Freelist_intf

type ('elt, 'blk_id, 't) version =

| For_blkids of {

e2b : 'elt -> 'blk_id;

b2e : 'blk_id -> 'elt;

}

| For_arbitrary_elts of {

alloc : unit -> ('blk_id, 't) Tjr_monad.m;

free : 'blk_id -> (unit, 't) Tjr_monad.m;

}

There are two versions of the freelist:

For_blkids: to implement the "usual" blk freelist, managing on-disk blocks; the elts are actually blk_ids

For_arbitrary_elts: which allocates and frees arbitrary elements; some other freelist of the first type provides the alloc and free blk functionality

type ('a, 'blk_id) fl_root_info = {

hd : 'blk_id;

tl : 'blk_id;

blk_len : int;

min_free : 'a option;

}

type ('a, 'blk_id, 't) fl_root_ops = {

write_root : ('a, 'blk_id) fl_root_info -> (unit, 't) Tjr_monad.m;

sync : unit -> (unit, 't) Tjr_monad.m;

}

write_root: we can recover the freelist providing we know the hd, tl and blk_len of the underlying plist

sync: ensure the freelist roots are actually on disk; redundant if we assume the blk_dev_ops write direct to disk with no cache; this sync is invoked by freelist_ops.sync

type sync_type = [

| `Tl_only

| `Tl_and_root_block

]

`Tl: this just ensures that the tl block of the freelist is synced; it does not update the root block

`Tl_and_root_block: sync tl and global root block

type fIXME

type ('blk_id, 't) freelist_ops = {

alloc : unit -> ('blk_id, 't) Tjr_monad.m;

alloc_many : int -> (fIXME, 't) Tjr_monad.m;

free : 'blk_id -> (unit, 't) Tjr_monad.m;

free_many : fIXME -> (unit, 't) Tjr_monad.m;

sync : sync_type -> (unit, 't) Tjr_monad.m;

}

alloc_many: int is the number of blk_ids (although this sort-of forces us to unmarshal the whole block; perhaps prefer also storing the number of elements in the block, precisely for this reason FIXME

free_many: free an entire list of blk_ids, by appending to this freelist

type 'a min_free_ops = {

min_free_alloc : 'a -> int -> 'a list * 'a;

}

Working with a min_free element, we provide a method that takes the min_free element, and a count of the number of frees required, and returns a list of free elements (of the same length as requested) and the new min_free

type 'a freelist = {

transient : 'a list;

min_free : ('a * 'a min_free_ops) option;

waiting : 'a Tjr_monad.event list;

disk_thread_active : bool;

}

The freelist state, in addition to the plist.

transient: a non-persistent list of free elts that can be allocated

min_free: free elts are stored in transient, or in the on-disk plist; usually we also have an "upper bound", beyond which we can still allocate elts freely (an elt is either in transient, allocated to some object, in the plist, or geq min_free)

waiting: a list of threads waiting for a disk process to return

NOTE: we assume that the state is accessed via with_state, ie, only one thread at a time

type ('a, 'blk_id) fl_root_mshlr = ('blk_id * 'blk_id * int * 'a option) Tjr_fs_shared.bp_mshlr

This type can be used to marshal the freelist root block

type ('a, 'buf, 'blk_id, 't) freelist_factory = < fl_root_mshlr : ('a, 'blk_id) fl_root_mshlr; from_disk : < version : ('a, 'blk_id, 't) version; fl_root_blk : 'blk_id; > -> < freelist_ops : ('blk_id, 't) freelist_ops; fl_root_ops : ('a, 'blk_id, 't) fl_root_ops; with_freelist : ('a freelist, 't) Tjr_monad.with_state; freelist_ref : 'a freelist Stdlib.ref; plist_ops : ('a, 'buf, 'blk_id, 't) Tjr_plist.Plist_intf.plist_ops; with_plist : (('blk_id, 'buf) Tjr_plist.Plist_intf.plist, 't) Tjr_monad.with_state; plist_ref : ('blk_id, 'buf) Tjr_plist.Plist_intf.plist Stdlib.ref; >; >

val make : < async : (unit -> (unit, 't) Tjr_monad.m) -> (unit, 't) Tjr_monad.m; event_ops : 't Tjr_monad.event_ops; monad_ops : 't Tjr_monad.monad_ops; plist : ('elt, 'buf, 'blk_id, 't) Tjr_plist.Plist_intf.plist_ops; root_block : ('elt, 'blk_id, 't) fl_root_ops; version : ('elt, 'blk_id, 't) version; with_freelist : ('elt freelist, 't) Tjr_monad.with_state; > -> ('elt, 't) freelist_ops

let make : 
< async : (unit -> (unit, 't) m) -> (unit, 't) m;
  event_ops : 't event_ops;
  monad_ops : 't monad_ops;
  plist : ('elt, 'buf, 'blk_id, 't) plist_ops;
  root_block : ('blk_id, 't) root_block_ops;
  version : ('elt, 'blk_id, 't) version;
  with_freelist : ('elt freelist, 't) with_state;
> -> ('elt, 't) freelist_ops
= make